Refactor dispatcher into different files
